requirements.txt
pandas
numpy
matplotlib
seaborn
joblib
scikit-learn
google-analytics-data
tensorflow
argparse
google-cloud-bigquery
Flask
flask-restful
flask-cors
requests